Prototype range scan operation
A range scan iterates over a [start key, end key] range.  It creates a
task that runs in the eleveldb thread pool and does not return until the
entire range has been visited.  It sends batches of data to a consumer
Erlang process.  The consume acks bytes as it receives the batches. The
producer task will block if the number of bytes without an ack reaches a
configurable limit. The consumer ack is mostly lockless, except when it
detects the consumer might have blocked after reaching the unacked max.

The hope is that this setup gives us an efficient range scan since using
iterators as currently implemented in eleveldb seems to be reducing
throughput by an order of magnitude or more. We suspect the constant
switching between Erlang and the async threads is the culprit, hence
this approach that has only two threads directly talking to each other
by passing large reference counted binaries or efficient ack operations.