Remove ack retries and add low watermark
The handling of the ack race where the producer had not yet gone to
sleep does not need the retry mechanism I had added. You can simply set
a flag to prevent it from going to sleep as that is all done in the same
critical section.

Add low watermark such that when the buffer gets full, we don't make the
producer spring into action as soon as enough bytes are acked to get it
under the max again. That produces many small I/O operations instead of
the larger sequential ones that we want.