Don't set filter_passed if range_start matches
In the case that the first key is not the range start key the code seeks
to the matching slice. But it also sets `filter_passed=false`. Later,
the iterator's new key is checked, as is `start_inclusive`. There is no
clause to set `filter_passed=true` though. I think that this is the
right way to handle `{true, true, true}={range_start, key_matches, iter_valid}`